requirements.txt
transformers
peft
accelerate
bitsandbytes
torch
sentence-transformers
faiss-cpu
gradio
pandas